support full scene updates mid-execution in PlanExecution
`plan.planning_scene_` was a trick to optionally apply a diff
on top of the scene maintained by `plan.planning_scene_monitor_`.

Everyday use does not rely on this additional diff.

This trick is severely flawed in multiple ways, e.g., changes to the `World`
of the monitored planning scene will not be noticed if a custom diff
is provided iff the diff includes changes to the world, because
`plan.planning_scene_` will provide its own world and the parent world
will never be looked at.

Crucially, if `plan.planning_scene_` was created as a `diff()` on top of
the scene of the monitor AND the monitor receives a full new scene AND
the monitor maintains diffs (the default in the `move_group` node),
`plan.planning_scene_.parent_` will point to `PlanningSceneMonitor::parent_scene_`
from then on and will thus not include the latest changes which are
maintained in `PlanningSceneMonitor::scene_`.
With the second full scene message, `plan.planning_scene_` will not get
any updates from the monitor anymore at all.

This patch changes the internal semantics of `plan.planning_scene_`
to override validation checks along the whole trajectory without
(deliberate) updates from the monitor.
On the other hand, if no diff is specified, the monitored scene will be
used everywhere, making changes in the monitored scene available to
validation checks everywhere.