Fix Pilz trajectory time stamps
The Pilz planner generated trajectories with identical subsequent time stamps,
which were rejected by the ROS controllers.

Pilz's appendWithStrictTimeIncrease() function had a single if-statement
that encompassed two conditions:

1. trajectory is empty
2. the end state of one segment differs from the start state of the next segment

In both cases, the whole trajectory segment was appended with a dt=0. However, that's correct only for case 1.
In case 2, we needed to apply a non-zero time offset dictated by the actual dt of that segment's first waypoint.