Simplify ros_control interface (#3707)
To enable ROS controllers to be used with MoveIt's controller manager, they need to be registered as "dummy" plugins of type JointTrajectoryControllerAllocator. However, this allocator just allocates a FollowJointTrajectoryControllerHandle. 
Instead of this complicated process, requiring additional controllers to be manually registered, we can just manage any ROS controller with a FollowJointTrajectory action interface.